Create identifier folder formove in all HandIdentifier classes
