--- src/main/java/jalau/at18/katas/pokerhands/adriana/HandIdentifier.java
@@ -1,7 +1,0 @@
-package jalau.at18.katas.pokerhands.adriana;
-
-abstract class HandIdentifier {
-
-    abstract boolean identify(PokerHand hand);
-    abstract HandKind getKind();
-}
