+++ src/main/java/jalau/at18/katas/pokerhands/adriana/identifier/FourOfAKindIdentifier.java
@@ -1,4 +1,6 @@
-package jalau.at18.katas.pokerhands.adriana;
+package jalau.at18.katas.pokerhands.adriana.identifier;
+
+import jalau.at18.katas.pokerhands.adriana.*;
 
 public class FourOfAKindIdentifier extends HandIdentifier {
     private static final int FOUR_OF_A_KIND = 4;
@@ -7,14 +9,14 @@ public class FourOfAKindIdentifier extends HandIdentifier {
     }
 
     @Override
-    boolean identify(PokerHand hand) {
+    public boolean identify(PokerHand hand) {
         CardValuesCounter counter = new CardValuesCounter(hand);
         CardValuesCount count = counter.getCount();
         return count.exists(FOUR_OF_A_KIND);
     }
 
     @Override
-    HandKind getKind() {
+    public HandKind getKind() {
         return HandKind.FOUR_EQUALS;
     }
 }
