+++ src/main/java/jalau/at18/katas/pokerhands/adriana/identifier/HandIdentifier.java
@@ -1,0 +1,9 @@
+package jalau.at18.katas.pokerhands.adriana.identifier;
+
+import jalau.at18.katas.pokerhands.adriana.*;
+
+public abstract class HandIdentifier {
+
+    public abstract boolean identify(PokerHand hand);
+    public abstract HandKind getKind();
+}
