+++ src/test/java/jalau/at18/katas/pokerhands/adriana/identifier/StraightFlushIdentifierTest.java
@@ -1,15 +1,17 @@
-package jalau.at18.katas.pokerhands.adriana;
+package jalau.at18.katas.pokerhands.adriana.identifier;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
-public class FlushIdentifierTest {
-    
+import jalau.at18.katas.pokerhands.adriana.*;
+
+public class StraightFlushIdentifierTest {
+
     @Test
-    public void shouldIdentifyFlushHand() {
-        FlushIdentifier identifier = new FlushIdentifier();
+    public void shouldIdentifyStraightFlushHand() {
+        StraightFlushIdentifier identifier = new StraightFlushIdentifier();
 
         PokerHand straightFlushHand = new PokerHand(new Card[]{
             new Card(CardValue.TWO, 'C'),
@@ -18,24 +20,24 @@ public class FlushIdentifierTest {
             new Card(CardValue.FIVE, 'C'),
             new Card(CardValue.SIX, 'C'),
         });
-        assertFalse(identifier.identify(straightFlushHand));
+        assertTrue(identifier.identify(straightFlushHand));
 
-        PokerHand FlushHand = new PokerHand(new Card[]{
+        PokerHand noConsecutiveValues = new PokerHand(new Card[]{
             new Card(CardValue.TWO, 'C'),
             new Card(CardValue.FOUR, 'C'),
             new Card(CardValue.FIVE, 'C'),
             new Card(CardValue.FIVE, 'C'),
             new Card(CardValue.SEVEN, 'C'),
         });
-        assertTrue(identifier.identify(FlushHand));
 
-        PokerHand straightHand = new PokerHand(new Card[]{
+        assertFalse(identifier.identify(noConsecutiveValues));
+        PokerHand noSameSuite = new PokerHand(new Card[]{
             new Card(CardValue.THREE, 'C'),
             new Card(CardValue.FOUR, 'C'),
             new Card(CardValue.FIVE, 'H'),
             new Card(CardValue.SIX, 'C'),
             new Card(CardValue.SEVEN, 'C'),
         });
-        assertFalse(identifier.identify(straightHand));
+        assertFalse(identifier.identify(noSameSuite));
     }
 }
